Finish Hover Animations for Follow Button
After many failed experiments, I was able to solve the problem: Initial
testing was showing that React only allowed the transition animations
to 'run' on a coponent that was there before and after the trigger. this
means that I had to keep the same component there. I was using
conditional rendering to swap components with a differnt heart icon URL.
The solution was to to use a single component and only modify the source
of the content attribute conditionally.