Fix bug when mouse entered from the side
The problem could indeed be fix by reverting to the original hover logic
that did not depend on the box data attribute. The problem was that
entering from the side triggered an onMouseEnter/Leave on the IMG and
not the div, and only the div had a data attribute, the problem was
solved by adding a 'white' box att to the images as well.