Finish Notification button Animations
I had not reaalized that the default (when not hovered over) notifications icon changed based onwhether it has been clicked on. I separated the click listener from the suscription button and added the new state, icon, click-logic amd hover effects. As with the other buttons, I had to add a data box-attribut to the icons as well, in order to prevent the lack of animation when the mouse entered from the side. Other than the sliding disappearing effect of the notification button, all button animations are now complete.