Begin effects for background IMG and Nav-buttons
The original BackgroundPic Component was deleted given all the styling
needed to be conditionally passed down to its image via the MainBar
component. I therefore decided to simplify and define it locally on the
Main Comp. There is still work to be done on the IMG sizing logic as
well as adding a proper slide animation to the nav buttons.