Add proto media query to navigation buttons
Given the complexity of the navigation bar, I had to keep the state
lifted at the main component so that it could be passed down via props
to both sets of navigation buttons. Upon mounting, both the main comp
and the Menu comp add an event listener to the window object that
updates their state, this allows the nav buttons to change based on
window resize. I still need to polish the onClick listener on the
drop-down menu so the order of the buttons is chamnged.