Fix Dropdown Menu Carousel and Selected functionality
Any hard-coded rendering of the navbuttons are now set
dynamically to their respective index of the 'pages' state array. A call
to the handleResizeMain() method was added on component did mount so
that the windowWidth state was right upon first loading the page. The
logic to handle a click on the dropdown was polished to correctly swap
items on the array.