Add basic (un-finished) Subscription Modal
I added a pop-up modal when clicking the subscribe button. It is far
from ready, but at least the emojis loaded from the server are there.
For a very strange reason, the emoji render lib randomly adds a
male/female emoji attached to some emojis, this throws of the alignment
and the emoji count.