Fix bug with male/female emoji
The emoji Render Lib was creating seamingly random inner text on the
emoji <span> the text had either a male or female emoji which threw off
the emoji count and the styling of the emoji container. I was able to
fix it by giving every parent node an id, and upon the component
mounting, iterate over all of them and remove that last text child if
the child count was more than 1.