Add mongoose disconnect and Docker Files
I added a line that diconnects from mongoose after the seeder script is
done filling the DB, it was necessary because the final 'npm start'
commamnd on docker-compose.yml would never run because the seeding file
would never successfully end. The Docker files allow the building of an
image which contains both an instance of Mongo and the Server as a
complete service.